requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
PyJWT==2.8.1
bcrypt==4.1.1
python-multipart==0.0.6
pandas==2.1.3